% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csquare.R
\docType{class}
\name{csquare}
\alias{csquare}
\title{csquare}
\format{\code{\link{R6Class}} object.}
\usage{
csquare
}
\value{
Object of \code{\link{R6Class}} for modelling a csquare
}
\description{
csquare
}
\section{Methods}{

\describe{
 \item{\code{new(str)}}{
   This method is used to instantiate an csquare
 }
}
}

\examples{
  csq <- csquare$new()
  

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{csquare}
