% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csquare4R.R
\docType{package}
\name{csquare4R}
\alias{csquare4R}
\alias{csquare4R-package}
\title{Function to encode/decode c-squares}
\description{
Function to encode/decode c-squares and bindings
for spatial objects.
}
\details{
\tabular{ll}{
   Package: \tab csquare4R\cr
   Type: \tab Package\cr
   Version: \tab 0.1\cr
   Date: \tab 2019-06-13\cr
   License: \tab MIT\cr
   LazyLoad: \tab yes\cr
 }
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
